+++ frontend/src/UserService.js
@@ -1,13 +1,13 @@
 import React, { Component } from 'react'
 
-export default class SignUp extends Component {
-
-
+export class UserService extends Component {
   render() {
     return (
       <div>
-        <h1>Sign up below:</h1>
+        
       </div>
     )
   }
 }
+
+export default UserService
