--- frontend/src/components/Login.js
@@ -1,12 +1,0 @@
-import React, { Component } from 'react'
-
-export default class Login extends Component {
-
-  render() {
-    return (
-      <div>
-        <h1>Login below:</h1>
-      </div>
-    )
-  }
-}
