+++ frontend/src/components/common/FooterBar.js
@@ -1,0 +1,10 @@
+import React from "react";
+import { Layout } from "antd";
+const { Footer: AntFtr } = Layout;
+
+const Footer = () => (
+    <AntFtr>
+        <div style={{ textAlign: "center" }}>Footer, etc</div>
+    </AntFtr>
+);
+export default Footer;
