+++ frontend/src/components/home/CardView.js
