+++ frontend/src/components/static/About.js
@@ -1,0 +1,8 @@
+import React from "react";
+
+const About = () => (
+    <div>
+        <h1>About</h1>
+    </div>
+);
+export default About;
