+++ frontend/src/components/user/Login.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Login = () => (
+    <div>
+        <h1>Login below:</h1>
+    </div>
+);
+
+export default Login;
