+++ frontend/src/components/user/SignUp.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const SignUp = () => (
+    <div>
+        <h1>Sign up below:</h1>
+    </div>
+);
+
+export default SignUp;
